\name{Beta.m}

\alias{Beta.m}

\docType{data}

\title{
   A simulated Beta matrix
}

\description{
  An example data set of a simulated 200x200 Beta matrix. It consists of 10 row block groups each containing 20 rows and 10 column blocks each containing 20 columns. 
}

\usage{data(Beta.m)}

\format{
  The format is:
  a matrix of dimension 200x200.
}

\references{
    Y. Li, B. Nan and J. Zhu (2015)
    Multivariate sparse group lasso for the multivariate multiple linear regression with an arbitrary group structure. Biometrics. DOI: 10.1111/biom.12292
}

\author{Yanming Li, Bin Nan, Ji Zhu}

\examples{
data(Beta.m)
}

\keyword{datasets}
